reworking deployment and operations (#141)
In some recent discussions, it became clear that users and customers
want specific direction on what instances, storage etc. to use

At the moment, our docs are fairly generic - have x amount of IOPS, use
n cores. But not so specific on actual instance, storage choices.

This PR aims to create a golden thread for users of specific Cloud
offerings to quickly get to a template deployment and get running.

---------

Co-authored-by: goodroot <9484709+goodroot@users.noreply.github.com>