Update partitions.md (Time Partitions > Properties > Hour) (#277)
Hello team!
I was checking the documentation about
[partitioning](https://questdb.com/docs/concept/partitions/#properties)
and came across a mismatch between the docs and my QuestDB instance
(9.1.1).

In short: the partition format for hourly partioned tables is not
`YYYY-MM-DD-HH` as specified in the docs but turns out to be
`YYYY-MM-DDTHH`. The UI running on port 9000 also confirms this:

<img width="627" height="178" alt="afbeelding"
src="https://github.com/user-attachments/assets/629d702d-d0e5-43a8-a571-a06fde80ad2e"
/>

Also, on the disk, in my quest db root folder where the database files
reside, the same format is used (so with a `T`).

Co-authored-by: Nick Woolmer <29717167+nwoolmer@users.noreply.github.com>